// blockchain/typechain-types/POSBusAuditLedger.ts
export interface POSBusAuditLedgerInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "getAudit"
      | "owner"
      | "storeAuditHash"
      | "transferOwnership"
      | "verifyAuditHash"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "AuditStored" | "OwnerTransferred"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "getAudit",
    values: [string, string]
  ): string;
  encodeFunctionData(functionFragment: "owner", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "storeAuditHash",
    values: [string, string, string, string]
  ): string;
  encodeFunctionData(
    functionFragment: "transferOwnership",
    values: [AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "verifyAuditHash",
    values: [string, string, string]
  ): string;

  decodeFunctionResult(functionFragment: "getAudit", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "owner", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "storeAuditHash",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "transferOwnership",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "verifyAuditHash",
    data: BytesLike
  ): Result;
}

export namespace AuditStoredEvent {
  export type InputTuple = [
    recordType: string,
    recordId: string,
    recordHash: string,
    metadataHash: string,
    timestamp: BigNumberish,
    anchorSender: AddressLike
  ];
  export type OutputTuple = [
    recordType: string,
    recordId: string,
    recordHash: string,
    metadataHash: string,
    timestamp: bigint,
    anchorSender: string
  ];
  export interface OutputObject {
    recordType: string;
    recordId: string;
    recordHash: string;
    metadataHash: string;
    timestamp: bigint;
    anchorSender: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace OwnerTransferredEvent {
  export type InputTuple = [previousOwner: AddressLike, newOwner: AddressLike];
  export type OutputTuple = [previousOwner: string, newOwner: string];
  export interface OutputObject {
    previousOwner: string;
    newOwner: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface POSBusAuditLedger extends BaseContract {
  connect(runner?: ContractRunner | null): POSBusAuditLedger;
  waitForDeployment(): Promise<this>;

  interface: POSBusAuditLedgerInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  getAudit: TypedContractMethod<
    [recordType: string, recordId: string],
    [
      [string, string, bigint, string] & {
        recordHash: string;
        metadataHash: string;
        blockTimestamp: bigint;
        anchorSender: string;
      }
    ],
    "view"
  >;

  owner: TypedContractMethod<[], [string], "view">;

  storeAuditHash: TypedContractMethod<
    [
      recordType: string,
      recordId: string,
      recordHash: string,
      metadataHash: string
    ],
    [void],
    "nonpayable"
  >;

  transferOwnership: TypedContractMethod<
    [newOwner: AddressLike],
    [void],
    "nonpayable"
  >;

  verifyAuditHash: TypedContractMethod<
    [recordType: string, recordId: string, recordHash: string],
    [boolean],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "getAudit"
  ): TypedContractMethod<
    [recordType: string, recordId: string],
    [
      [string, string, bigint, string] & {
        recordHash: string;
        metadataHash: string;
        blockTimestamp: bigint;
        anchorSender: string;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "owner"
  ): TypedContractMethod<[], [string], "view">;
  getFunction(
    nameOrSignature: "storeAuditHash"
  ): TypedContractMethod<
    [
      recordType: string,
      recordId: string,
      recordHash: string,
      metadataHash: string
    ],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "transferOwnership"
  ): TypedContractMethod<[newOwner: AddressLike], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "verifyAuditHash"
  ): TypedContractMethod<
    [recordType: string, recordId: string, recordHash: string],
    [boolean],
    "view"
  >;

  getEvent(
    key: "AuditStored"
  ): TypedContractEvent<
    AuditStoredEvent.InputTuple,
    AuditStoredEvent.OutputTuple,
    AuditStoredEvent.OutputObject
  >;
  getEvent(
    key: "OwnerTransferred"
  ): TypedContractEvent<
    OwnerTransferredEvent.InputTuple,
    OwnerTransferredEvent.OutputTuple,
    OwnerTransferredEvent.OutputObject
  >;

  filters: {
    "AuditStored(string,string,string,string,uint256,address)": TypedContractEvent<
      AuditStoredEvent.InputTuple,
      AuditStoredEvent.OutputTuple,
      AuditStoredEvent.OutputObject
    >;
    AuditStored: TypedContractEvent<
      AuditStoredEvent.InputTuple,
      AuditStoredEvent.OutputTuple,
      AuditStoredEvent.OutputObject
    >;

    "OwnerTransferred(address,address)": TypedContractEvent<
      OwnerTransferredEvent.InputTuple,
      OwnerTransferredEvent.OutputTuple,
      OwnerTransferredEvent.OutputObject
    >;
    OwnerTransferred: TypedContractEvent<
      OwnerTransferredEvent.InputTuple,
      OwnerTransferredEvent.OutputTuple,
      OwnerTransferredEvent.OutputObject
    >;
  };
}